added authorization

+++ routes/delete-routes.go
@@ -2,16 +2,23 @@ package routes
 
 import (
 	"api/controllers"
+	"api/models"
+	"api/utils"
 	"database/sql"
 	"net/http"
 
 	"github.com/gorilla/mux"
 )
 
-func DeleteMagazine(db *sql.DB) http.HandlerFunc {
+func DeleteMagazine(db *sql.DB, authDB []models.Authorize) http.HandlerFunc {
 	return func(rw http.ResponseWriter, r *http.Request) {
 		params := mux.Vars(r)
 		id := params["compID"]
-		controllers.DeleteMagazineCT(db, rw, id)
+		mainauS := utils.AuthorizeMethod(r, authDB)
+		if mainauS == authDB[0].MainAuth {
+			controllers.DeleteMagazineCT(db, rw, id)
+		} else {
+			utils.JsonResponse("Bad token!", false, rw)
+		}
 	}
 }
